JWTAuthorization 구현

+++ src/main/java/com/example/GoldenReport/Service/JWT/JWTType.java
@@ -9,5 +9,5 @@ public enum JWTType {
     ACCESS (3600),
     REFRESH (3600 * 24 * 3);
 
-    private final long validTime;
+    private final int validTime;
 }
